pyproject.toml
[project]
name = "nse-am"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "pydantic>=2.5,<3",
    "PyYAML>=6.0",
]
